predelana struktura PlayerData a pridana trida Player; zprovoznen multiplayer; pridany helper metody pro zapis a cteni objektu do zprav

+++ Application/src/Core/Scene/Entities/IRotable.cs
@@ -5,6 +5,6 @@ namespace Orbit.Core.Scene.Entities
 {
     public interface IRotable
     {
-        int Rotation { get; set; }
+        float Rotation { get; set; }
     }
 }
